third_party/blink/web_tests/external/wpt/html/semantics/embedded-content/the-img-element/relevant-mutations-expected.txt

This is a testharness.js-based test.
Found 113 tests; 106 PASS, 7 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS src set
PASS src changed
PASS src removed
PASS srcset set
PASS srcset changed
PASS srcset removed
PASS sizes set
PASS sizes changed
PASS sizes removed
PASS src set to same value
PASS crossorigin absent to empty, src absent
PASS crossorigin absent to anonymous, src absent
PASS crossorigin absent to use-credentials, src absent
PASS crossorigin empty to absent, src absent
PASS crossorigin empty to use-credentials, src absent
PASS crossorigin anonymous to absent, src absent
PASS crossorigin anonymous to use-credentials, src absent
PASS crossorigin use-credentials to absent, src absent
PASS crossorigin use-credentials to empty, src absent
PASS crossorigin use-credentials to anonymous, src absent
PASS crossorigin use-credentials to invalid, src absent
PASS crossorigin absent to empty, src already set
PASS crossorigin absent to anonymous, src already set
PASS crossorigin absent to use-credentials, src already set
PASS crossorigin empty to absent, src already set
PASS crossorigin empty to use-credentials, src already set
PASS crossorigin anonymous to absent, src already set
PASS crossorigin anonymous to use-credentials, src already set
PASS crossorigin use-credentials to absent, src already set
PASS crossorigin use-credentials to empty, src already set
PASS crossorigin use-credentials to anonymous, src already set
PASS crossorigin use-credentials to invalid, src already set
PASS referrerpolicy absent to no-referrer-when-downgrade, src absent
PASS referrerpolicy absent to no-referrer, src absent
PASS referrerpolicy empty to no-referrer-when-downgrade, src absent
PASS referrerpolicy empty to no-referrer, src absent
PASS referrerpolicy no-referrer-when-downgrade to absent, src absent
PASS referrerpolicy no-referrer-when-downgrade to empty, src absent
PASS referrerpolicy no-referrer-when-downgrade to no-referrer, src absent
PASS referrerpolicy no-referrer-when-downgrade to invalid, src absent
PASS referrerpolicy no-referrer to absent, src absent
PASS referrerpolicy no-referrer to empty, src absent
PASS referrerpolicy no-referrer to no-referrer-when-downgrade, src absent
PASS referrerpolicy no-referrer to invalid, src absent
PASS referrerpolicy absent to no-referrer-when-downgrade, src already set
PASS referrerpolicy absent to no-referrer, src already set
PASS referrerpolicy empty to no-referrer-when-downgrade, src already set
PASS referrerpolicy empty to no-referrer, src already set
PASS referrerpolicy no-referrer-when-downgrade to absent, src already set
PASS referrerpolicy no-referrer-when-downgrade to empty, src already set
PASS referrerpolicy no-referrer-when-downgrade to no-referrer, src already set
PASS referrerpolicy no-referrer-when-downgrade to invalid, src already set
PASS referrerpolicy no-referrer to absent, src already set
PASS referrerpolicy no-referrer to empty, src already set
PASS referrerpolicy no-referrer to no-referrer-when-downgrade, src already set
PASS referrerpolicy no-referrer to invalid, src already set
FAIL inserted into picture assert_unreached: update the image data didn't run Reached unreachable code
FAIL removed from picture assert_unreached: update the image data didn't run Reached unreachable code
PASS parent is picture, previous source inserted
PASS parent is picture, previous source removed
PASS parent is picture, previous source has srcset set
PASS parent is picture, previous source has srcset changed
PASS parent is picture, previous source has srcset removed
PASS parent is picture, previous source has sizes set
PASS parent is picture, previous source has sizes changed
PASS parent is picture, previous source has sizes removed
PASS parent is picture, previous source has media set
PASS parent is picture, previous source has media changed
PASS parent is picture, previous source has media removed
PASS parent is picture, previous source has type set
PASS parent is picture, previous source has type changed
PASS parent is picture, previous source has type removed
FAIL srcset is set to same value assert_unreached: update the image data was run Reached unreachable code
FAIL sizes is set to same value assert_unreached: update the image data was run Reached unreachable code
PASS crossorigin state not changed: absent, removeAttribute
PASS crossorigin state not changed: empty to anonymous
PASS crossorigin state not changed: anonymous to foobar
PASS crossorigin state not changed: use-credentials to USE-CREDENTIALS
PASS referrerpolicy state not changed: absent, removeAttribute
PASS referrerpolicy state not changed: empty to empty
PASS referrerpolicy state not changed: empty to invalid
PASS referrerpolicy state not changed: absent to invalid
PASS referrerpolicy state not changed: no-referrer to NO-REFERRER
PASS referrerpolicy state not changed: invalid to other-invalid
PASS inserted into picture ancestor
PASS removed from picture ancestor
PASS ancestor picture has a source inserted
PASS ancestor picture has a source removed
PASS ancestor picture; previous sibling source inserted
PASS ancestor picture; previous sibling source removed
FAIL parent is picture, following sibling source inserted assert_unreached: update the image data was run Reached unreachable code
FAIL parent is picture, following sibling source removed assert_unreached: update the image data was run Reached unreachable code
FAIL parent is picture, following sibling source has srcset set assert_unreached: update the image data was run Reached unreachable code
PASS media on img set
PASS type on img set
PASS class on img set
PASS alt on img set
PASS src on previous sibling source set
PASS class on previous sibling source set
PASS inserted/removed children of img
PASS picture is inserted; img has src
PASS picture is inserted; img has srcset
PASS picture is inserted; img has previous sibling source
PASS picture is inserted; img has following sibling source
PASS picture is removed; img has src
PASS picture is removed; img has srcset
PASS picture is removed; img has previous sibling source
PASS picture is removed; img has following sibling source
PASS parent is picture, following img inserted
PASS parent is picture, following img removed
PASS parent is picture, following img has src set
PASS parent is picture, following img has srcset set
PASS parent is picture, following img has sizes set
Harness: the test ran to completion.
